// frontend/src/utils/types/Action.d.ts
import { InformationTexts } from "../OfferView/InformationTexts";

export type Action = {
	type: string;
	newSite: string;
	newProductType: string;
	newInformationTexts: InformationTexts;
	newImageUrl: string;
	newPercent: number;
	newMinimumAmount: number;
	newMaximumAmount: number;
	newScale: string;
	newCurrentRegisterPage: string;
	newSelectedAccountTypeIndex: number;
	newEmail: string;
	newPassword: string;
	newName: string;
	newSurname: string;
	newPhoneNumber: string;
	newDocumentType: string;
	newDocumentId: string;
	newPesel: string;
	newBirthday: string;
	newMotherSurname: string;
	newCountry: string;
	newPostCode: string;
	newCity: string;
	newHouseNumber: string;
	newApartmentNumber: string;
	newAccountType: string;
	newCurrency: string;
	newStreet: string;
	newIsLoggedIn: boolean;
	newAccountsArray: Array;
	newDepositsArray: Array;
	newTransactionsArray: Array;
	newDailyLimit: number;
	newSingleTransactionLimit: number;
	newAccountId: string;
	newAmount: number;
	newDepositType: string;
	newDepositEndDate: string;
	newDepositNumber: string;
	newDepositPercent: string;
	newDepositMoney: number;
	newDepositCurrency: string;
	newTitle: string;
	newReceiverAccountObject: {
		accountId: string;
		accountNumber: string;
	};
	newType: string;
};